+++ components/checkbox/checkbox.vue
@@ -69,7 +69,31 @@ export default {
 
   inheritAttrs: false,
 
-  emits: ['input', 'focusin', 'focusout'],
+  emits: [
+    /**
+     * Native input event
+     *
+     * @event input
+     * @type {Boolean}
+     */
+    'input',
+
+    /**
+     * Native input focusin event
+     *
+     * @event focusin
+     * @type {FocusEvent}
+     */
+    'focusin',
+
+    /**
+     * Native input focusout event
+     *
+     * @event focusout
+     * @type {FocusEvent}
+     */
+    'focusout',
+  ],
 
   computed: {
     inputValidationClass () {
@@ -98,7 +122,6 @@ export default {
             Check if any usages of this component leverage $listeners and either remove if unused or scope the removal
             and migration prior to upgrading to Vue 3.x
         */
-        // eslint-disable-next-line vue/no-deprecated-dollar-listeners-api
         ...this.$listeners,
         /*
          * Override input listener to as no-op. Prevents parent input listeners from being passed through onto the input