+++ components/combobox/combobox.vue
@@ -22,12 +22,24 @@
       @focusout="clearHighlightIndex"
       @mousemove.capture="onMouseHighlight"
     >
+      <combobox-loading-list
+        v-if="isLoading && !listRenderedOutside"
+        v-bind="listProps"
+      />
+      <combobox-empty-list
+        v-else-if="isListEmpty && !listRenderedOutside"
+        v-bind="listProps"
+        :message="emptyStateMessage"
+      />
       <!-- @slot Slot for the combobox list element -->
       <slot
+        v-else
         name="list"
         :list-props="listProps"
         :opened="onOpen"
         :clear-highlight-index="clearHighlightIndex"
+        :is-loading="isLoading"
+        :is-list-empty="isListEmpty"
       />
     </div>
   </div>
@@ -36,9 +48,16 @@
 <script>
 import KeyboardNavigation from '@/common/mixins/keyboard_list_navigation';
 import { getUniqueString } from '@/common/utils';
+import ComboboxLoadingList from './combobox_loading-list.vue';
+import ComboboxEmptyList from './combobox_empty-list.vue';
 
 export default {
   name: 'DtCombobox',
+
+  components: {
+    ComboboxLoadingList,
+    ComboboxEmptyList,
+  },
 
   mixins: [
     KeyboardNavigation({
@@ -101,9 +120,56 @@ export default {
       type: Boolean,
       default: false,
     },
+
+    /**
+     * Determines when to show the skeletons and also controls aria-busy attribute.
+     */
+    loading: {
+      type: Boolean,
+      default: false,
+    },
+
+    /**
+     * Message to show when the list is empty
+     */
+    emptyStateMessage: {
+      type: String,
+      default: '',
+    },
   },
 
-  emits: ['select', 'escape', 'highlight', 'opened'],
+  emits: [
+    /**
+     * Event fired when item selected
+     *
+     * @event select
+     * @type {Number}
+     */
+    'select',
+
+    /**
+     * Event fired when pressing escape
+     *
+     * @event escape
+     */
+    'escape',
+
+    /**
+     * Event fired when the highlight changes
+     *
+     * @event highlight
+     * @type {Number}
+     */
+    'highlight',
+
+    /**
+     * Event fired when list is shown or hidden
+     *
+     * @event opened
+     * @type {Boolean}
+     */
+    'opened',
+  ],
 
   data () {
     return {
@@ -111,6 +177,8 @@ export default {
       // of this component, this is the ref to that dom element. Set
       // by the onOpen method.
       outsideRenderedListRef: null,
+      isListEmpty: undefined,
+      isLoading: undefined,
     };
   },
 
@@ -146,13 +214,14 @@ export default {
     },
 
     activeItemId () {
-      if (!this.showList || this.highlightIndex < 0) {
+      if (!this.showList || this.highlightIndex < 0 || this.loading) {
         return;
       }
       return this.highlightId;
     },
 
     activeItemEl () {
+      if (!this.highlightId) return '';
       return this.getListElement().querySelector('#' + this.highlightId);
     },
   },
@@ -160,22 +229,39 @@ export default {
   watch: {
     showList (showList) {
       // When the list's visibility changes reset the highlight index.
-      this.$nextTick(function () {
-        if (!this.listRenderedOutside) {
-          this.setInitialHighlightIndex();
-          this.$emit('opened', showList);
-        }
-      });
+
+      if (!this.listRenderedOutside) {
+        this.setInitialHighlightIndex();
+        this.$emit('opened', showList);
+      }
 
       if (!showList && this.outsideRenderedListRef) {
         this.outsideRenderedListRef.removeEventListener('mousemove', this.onMouseHighlight);
         this.outsideRenderedListRef = null;
+        this.isListEmpty = undefined;
       }
     },
+
+    loading (isLoading) {
+      this.isListEmpty = undefined;
+      this.isLoading = isLoading;
+      this.$nextTick(() => {
+        this.isListEmpty = this.checkItemsLength();
+        this.setInitialHighlightIndex();
+      });
+    },
+  },
+
+  async mounted () {
+    this.isLoading = this.loading;
+    await this.$nextTick();
+    this.isListEmpty = this.checkItemsLength();
   },
 
   methods: {
     onMouseHighlight (e) {
+      if (this.loading) return;
+
       const liElement = e.target.closest('li');
 
       if (liElement && this.highlightId !== liElement.id) {
@@ -184,7 +270,7 @@ export default {
     },
 
     getListElement () {
-      return this.outsideRenderedListRef ?? this.$refs.listWrapper.querySelector(`#${this.listId}`);
+      return this.outsideRenderedListRef ?? this.$refs.listWrapper?.querySelector(`#${this.listId}`);
     },
 
     clearHighlightIndex () {
@@ -194,10 +280,13 @@ export default {
     },
 
     afterHighlight () {
+      if (this.loading) return;
       this.$emit('highlight', this.highlightIndex);
     },
 
     onEnterKey () {
+      if (this.loading || this.isListEmpty) return;
+
       if (this.highlightIndex >= 0) {
         this.$emit('select', this.highlightIndex);
       }
@@ -213,21 +302,31 @@ export default {
       this.$emit('opened', open);
 
       if (open) {
+        this.isListEmpty = this.checkItemsLength();
         this.setInitialHighlightIndex();
       }
     },
 
     onKeyValidation (e, eventHandler) {
-      if (!this.showList || !this.getListElement()) { return; }
+      if (!this.showList || !this.getListElement()) return;
 
       this[eventHandler](e);
     },
 
     setInitialHighlightIndex () {
-      if (this.showList) {
-        // When the list's is shown, reset the highlight index.
-        this.setHighlightIndex(0);
-      }
+      if (!this.showList) return;
+      this.$nextTick(() => {
+      // When the list's is shown, reset the highlight index.
+      // If the list is loading, set to -1
+        this.setHighlightIndex(this.loading ? -1 : 0);
+      });
+    },
+
+    checkItemsLength () {
+      if (!this.showList) return undefined;
+      const list = this.getListElement();
+      const options = list?.querySelectorAll(`[role="option"]`);
+      return options?.length === 0;
     },
   },
 };