+++ components/input/input.vue
@@ -13,6 +13,7 @@
       <slot name="labelSlot">
         <div
           v-if="label"
+          ref="label"
           data-qa="dt-input-label"
           :class="[
             'base-input__label-text',
@@ -26,6 +27,7 @@
       <div
         v-if="$slots.description || description || shouldValidateLength"
         :id="descriptionKey"
+        ref="description"
         :class="[
           'base-input__description',
           'd-description',
@@ -114,6 +116,9 @@ import {
 import { DtValidationMessages } from '../validation_messages';
 import { MessagesMixin } from '@/common/mixins/input.js';
 
+/**
+ * @property {Boolean} placeholder attribute
+ */
 export default {
   name: 'DtInput',
 
@@ -226,7 +231,70 @@ export default {
     },
   },
 
-  emits: ['blur', 'input', 'clear', 'focus', 'focusin', 'focusout', 'update:length', 'update:invalid'],
+  emits: [
+    /**
+     * Native input event
+     *
+     * @event input
+     * @type {String}
+     */
+    'input',
+
+    /**
+     * Native input blur event
+     *
+     * @event blur
+     * @type {FocusEvent}
+     */
+    'blur',
+
+    /**
+     * Input clear event
+     *
+     * @event clear
+     */
+    'clear',
+
+    /**
+     * Native input focus event
+     *
+     * @event focus
+     * @type {FocusEvent}
+     */
+    'focus',
+
+    /**
+     * Native input focusin event
+     *
+     * @event focusin
+     * @type {FocusEvent}
+     */
+    'focusin',
+
+    /**
+     * Native input focusout event
+     *
+     * @event focusout
+     * @type {FocusEvent}
+     */
+    'focusout',
+
+    /**
+     * Length of the input when currentLength prop is not passed
+     *
+     * @event update:length
+     * @type {Number}
+     */
+    'update:length',
+
+    /**
+     * Result of the input validation
+     *
+     * @event update:invalid
+     * @type {Boolean}
+     */
+    'update:invalid',
+  ],
 
   data () {
     return {