+++ components/list_item/list_item.vue
@@ -96,7 +96,39 @@ export default {
     },
   },
 
-  emits: ['click', 'keydown', 'mousemove', 'mouseleave'],
+  emits: [
+    /**
+     * Native click event
+     *
+     * @event click
+     * @type {PointerEvent | KeyboardEvent}
+     */
+    'click',
+
+    /**
+     * Key down event
+     *
+     * @event keydown
+     * @type {KeyboardEvent}
+     */
+    'keydown',
+
+    /**
+     * Native mouse move event
+     *
+     * @event mousemove
+     * @type {MouseEvent}
+     */
+    'mousemove',
+
+    /**
+     * Native mouse leave event
+     *
+     * @event mouseleave
+     * @type {MouseEvent}
+     */
+    'mouseleave',
+  ],
 
   data () {
     return {
@@ -123,6 +155,13 @@ export default {
             this.onClick(event);
           }
           this.$emit('keydown', event);
+        },
+
+        mousedown: event => {
+          if (!this.isFocusable) {
+            event.preventDefault();
+          }
+          this.$emit('mousedown', event);
         },
 
         mousemove: event => {