+++ components/toast/toast.vue
@@ -170,7 +170,15 @@ export default {
     },
   },
 
-  emits: ['close'],
+  emits: [
+    /**
+     * Close button click event
+     *
+     * @event close
+     */
+    'close',
+    'update:show',
+  ],
 
   data () {
     return {
@@ -194,7 +202,6 @@ export default {
 
     noticeActionListeners () {
       return {
-        // eslint-disable-next-line vue/no-deprecated-dollar-listeners-api
         ...this.$listeners,
 
         close: event => {
@@ -225,7 +232,6 @@ export default {
     },
   },
 
-  // eslint-disable-next-line vue/no-deprecated-destroyed-lifecycle
   destroyed () {
     if (this.shouldSetTimeout) {
       clearTimeout(this.displayTimer);