+++ components/tooltip/tooltip_default.story.vue
@@ -1,4 +1,3 @@
-<!-- eslint-disable vue/no-deprecated-v-bind-sync -->
 <template>
   <div
     :class="[
@@ -16,9 +15,10 @@
         :message="message"
         :fallback-placements="fallbackPlacements"
         :offset="offset"
+        :sticky="sticky"
         :content-class="contentClass"
         :transition="transition"
-        :show.sync="show"
+        :show="showTooltip"
         @shown="onShown"
       >
         <template
@@ -55,6 +55,14 @@ export default {
   computed: {
     buttonKind () {
       return this.inverted ? 'inverted' : 'default';
+    },
+
+    showTooltip () {
+      if (this.globalShow) {
+        return true;
+      }
+
+      return this.show;
     },
   },
 };