+++ index.js
@@ -8,7 +8,6 @@ export * from './components/button_group';
 export * from './components/combobox';
 export * from './components/collapsible';
 export * from './components/dropdown';
-export * from './components/emoji';
 export * from './components/input';
 export * from './components/input_group';
 export * from './components/modal';
@@ -17,6 +16,7 @@ export * from './components/list_section';
 export * from './components/list_item';
 export * from './components/link';
 export * from './components/notice';
+export * from './components/pagination';
 export * from './components/popover';
 export * from './components/radio';
 export * from './components/radio_group';
@@ -35,14 +35,12 @@ export * from './components/root_layout';
 
 /// Recipes
 export * from './recipes/comboboxes/combobox_with_popover';
+export * from './recipes/comboboxes/combobox_multi_select';
 export * from './recipes/list_items/contact_info';
 export * from './recipes/notices/top_banner_info';
 
 // Mixins
 export * from './common/mixins';
-
-// Emoji
-export * from './common/emoji';
 
 // Constants
 export {