+++ recipes/comboboxes/combobox_with_popover/combobox_with_popover_default.story.vue
@@ -1,6 +1,8 @@
 <template>
   <dt-recipe-combobox-with-popover
     ref="comboboxWithPopover"
+    :empty-state-message="emptyStateMessage"
+    :loading="loading"
     :list-aria-label="listAriaLabel"
     :max-height="maxHeight"
     :max-width="maxWidth"
@@ -10,6 +12,7 @@
     :list-id="listId"
     :list-class="listClass"
     :open-with-arrow-keys="openWithArrowKeys"
+    :has-suggestion-list="hasSuggestionList"
     @escape="onComboboxEscape"
     @highlight="onHighlight"
     @select="onComboboxSelect"
@@ -22,11 +25,12 @@
       <span v-html="header" />
     </template>
 
-    <template #input="{ inputProps }">
+    <template #input="{ inputProps, onInput }">
       <dt-input
         v-model="value"
         placeholder="Select one or start typing"
         v-bind="inputProps"
+        @input="onInput"
       />
     </template>
     <template #list="{ listProps }">
@@ -81,20 +85,6 @@ export default {
   data () {
     return {
       value: '',
-      items: [
-        { id: 'item1', number: '(732) 338-2720', type: 'MAINLINE' },
-        { id: 'item2', number: '(732) 338-2722', type: 'MAINLINE' },
-        { id: 'item3', number: '(732) 338-2725', type: 'MAINLINE' },
-        { id: 'item4', number: '(732) 338-2764', type: 'MAINLINE' },
-        { id: 'item5', number: '(732) 338-2784', type: 'MAINLINE' },
-        { id: 'item6', number: '(732) 338-2743', type: 'MAINLINE' },
-        { id: 'item7', number: '(732) 338-2728', type: 'MAINLINE' },
-        { id: 'item8', number: '(732) 338-2769', type: 'Other' },
-        { id: 'item9', number: '(732) 338-2723', type: 'MAINLINE' },
-        { id: 'item10', number: '(732) 338-2729', type: 'MAINLINE' },
-        { id: 'item11', number: '(732) 338-2489', type: 'MAINLINE' },
-        { id: 'item12', number: '(732) 338-2756', type: 'Other' },
-      ],
     };
   },